app/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.2
decorator==4.4.2
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-celery-beat==2.5.0
django-recaptcha==3.0.0
django-redis==5.3.0
django-timezone-field==5.1
donatello==1.0.7
frozenlist==1.4.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
kombu==5.3.1
moviepy==1.0.3
multidict==6.0.4
numpy==1.25.2
oauthlib==3.2.2
Pillow==10.0.0
proglog==0.1.10
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-magic==0.4.27
python3-openid==3.2.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
python-dotenv==1.1.0
gunicorn==23.0.0
